refactor config.py into seperate modules for maintainability
